Updating for script1
The main idea of this script is to compare two values from two servers. Only two.
The values can be selected by key: If we have a space in the resulting string, we do a split.
If there is no space, then the string is whole and you need to compare the whole string.